chore: fix next target 최신 js기능을 사용하기 위해 target을 es2015 스펙으로 결정
