+++ packages/front-end/tsconfig.json
@@ -12,6 +12,7 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
+    "target":"ES2015",
     "plugins": [
       {
         "name": "next"
